Create mock-model
This document exploits a few JavaScript 'facts of life' in an attempt to model the behind the scenes stuff that is going on in the standard-practice document in simpler, albeit more verbose, JavaScript terminology. There are a few snags though; this document is unable to successfully demonstrate the inheritance phenomenon.  My hope is that by repeatedly tweaking this model, inheritance proper behavior will emerge and I'll finally understand the beast that is inheritance-based object oriented programming.